// comunicacionesDirectas/src/shared/vendor/tracasa-real/components/chat/tarjeta-comunicacion/tarjeta-comunicacion.module.scss.d.ts
declare const styles: {
  tarjeta: string;
  tarjeta__seleccionada: string;
  tarjeta__archivada: string;
  tarjeta__disabled: string;
  tarjeta__boton: string;
  texto__titulo: string;
  texto__subtitulo: string;
  texto__fecha: string;
  texto__inferior: string;
  fila__superior: string;
  fila__inferior: string;
  icono: string;
  icono__primario: string;
  icono__secundario: string;
  badge: string;
};

export default styles;